1-sem/RAS/Trabalho/Parte3/PictuRAS-main/backend/images_app/requirements.txt
Flask==3.1.0
Flask_Cors==5.0.0
mongoengine==0.29.1
python-dotenv==1.0.1
Pillow==11.1.0